moved timestamped node updater to graph optimizer
